ubah jadi biginteger

+++ U-mad/migrations/20220421033042-create-user-detail.js
@@ -21,7 +21,7 @@ module.exports = {
         allowNull: false
       },
       phoneNumber: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.BIGINT(14),
         allowNull: false
       },
       UserId: {
